+++ src/typed_settings/_core.py
@@ -7,9 +7,9 @@ from typing import Any, Dict, Iterable, List, Optional, Sequence, Type, Union
 
 import attrs
 
-from ._dict_utils import _deep_options, _merge_dicts, _set_path
 from .attrs import METADATA_KEY
 from .converters import BaseConverter, default_converter, from_dict
+from .dict_utils import deep_options, merge_dicts, set_path
 from .loaders import EnvLoader, FileLoader, Loader, TomlFormat
 from .types import AUTO, ST, OptionList, SettingsClass, _Auto
 
@@ -178,7 +178,7 @@ def load(
     )
     settings = _load_settings(
         cls=cls,
-        options=_deep_options(cls),
+        options=deep_options(cls),
         loaders=loaders,
     )
 
@@ -213,7 +213,7 @@ def load_settings(
         converter = default_converter()
     settings = _load_settings(
         cls=cls,
-        options=_deep_options(cls),
+        options=deep_options(cls),
         loaders=loaders,
     )
     return from_dict(settings, cls, converter)
@@ -239,11 +239,11 @@ def _load_settings(
             continue
         if isinstance(opt.field.default, attrs.Factory):  # type: ignore
             continue
-        _set_path(settings, opt.path, opt.field.default)
+        set_path(settings, opt.path, opt.field.default)
 
     loaded_settings = [loader(cls, options) for loader in loaders]
 
     for ls in loaded_settings:
-        _merge_dicts(options, settings, ls)
+        merge_dicts(options, settings, ls)
 
     return settings