+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
-from typed_settings._dict_utils import _deep_options
 from typed_settings.attrs import option, settings
+from typed_settings.dict_utils import deep_options
 from typed_settings.types import OptionList
 
 
@@ -27,4 +27,4 @@ def settings_cls() -> type:
 
 @pytest.fixture
 def options(settings_cls: type) -> OptionList:
-    return _deep_options(settings_cls)
+    return deep_options(settings_cls)